[リリース]NyARToolkit 1.0.0
git-svn-id: https://svn.sourceforge.jp/svnroot/nyartoolkit/NyARToolkit/trunk@80 7cac0a50-4618-4814-88d0-24b83990f816

+++ src/jp/nyatla/nyartoolkit/core/match/NyARMatchPatt_BlackWhite.java
@@ -51,7 +51,7 @@ public class NyARMatchPatt_BlackWhite implements ARMatchPatt{
     {
 	width=i_target_patt.getWidth();
 	height=i_target_patt.getHeight();
-	short[][][] data=i_target_patt.getPatArray(); 	
+	int[][][] data=i_target_patt.getPatArray(); 	
 	input=new int[height][width][3];
  
         int sum = ave = 0;
